Use base handler for delete user job
